+++ models.h
@@ -14,11 +14,11 @@ namespace rate{
      * epsilon = equilibrium term of the anti-hebbian dynamics ***stores e/N
      */
     struct parameters{
-        std::vector<float>::size_type N;
-        float gamma;
-        float tau;
-        float eps_ah;
-        float eps_h;
+        std::vector<double>::size_type N;
+        double gamma;
+        double tau;
+        double eps_ah;
+        double eps_h;
     };
 
     /* Group the network vectors and matrix variables
@@ -34,13 +34,13 @@ namespace rate{
      * dynamics used to temporarily destabilise attractors created by Hebbian learning
      */
     struct network_var{
-        std::vector<float> s_vec;
-        std::vector<float> h_vec;
-        std::vector<float> f_vec;
+        std::vector<double> s_vec;
+        std::vector<double> h_vec;
+        std::vector<double> f_vec;
 
-        std::vector<std::vector<float>> P_matrix;
-        std::vector<std::vector<float>> Hebb_matrix;
-        std::vector<std::vector<float>> antiHebb_matrix;
+        std::vector<std::vector<double>> P_matrix;
+        std::vector<std::vector<double>> Hebb_matrix;
+        std::vector<std::vector<double>> antiHebb_matrix;
     };
 
     /* Group the variables responsible for the random number generators
@@ -50,7 +50,7 @@ namespace rate{
      */
     struct random_device{
         std::mt19937 eng; //random number engine. Needs a seed (random device)
-        std::uniform_real_distribution<float> dist; // distribution used for random process
+        std::uniform_real_distribution<double> dist; // distribution used for random process
     };
 
 
@@ -78,24 +78,24 @@ namespace rate{
             random_device device;
 
             // prepares the random device
-            void prepare_random_device(float a, float b);
+            void prepare_random_device(double a, double b);
 
             //Draws from the distribution
-            float draw();
+            double draw();
 
         public:
             // Constructor of the ItineratedMap class
             // Initiate the state vector with random values in the range (a, b)
-            ItineratedMap(parameters& par, float a, float b);
+            ItineratedMap(parameters& par, double a, double b);
 
             // Instanciate a class from parameters and initial state provided by the user
-            ItineratedMap(parameters& par, std::vector<float>& vec);
+            ItineratedMap(parameters& par, std::vector<double>& vec);
 
             // Stores a random pattern, where xi in [-1, 1].
-            void add_random_pattern(float prob, float strength);
+            void add_random_pattern(double prob, double strength);
 
             // Stores a pattern provided by the user
-            void add_exterior_pattern(std::vector<float>& vec, float strength);
+            void add_exterior_pattern(std::vector<double>& vec, double strength);
 
             // Generates the Hebbian matrix part of the connection
             void make_hebb_matrix();
@@ -103,23 +103,23 @@ namespace rate{
             /*---- Probing methods ----*/
             /*---- These methods return references, be careful! ----*/
             // Get pattern by index mu
-            std::vector<float>& get_xi(int u);
+            std::vector<double>& get_xi(int u);
 
             //Probe state vector;
-            std::vector<float>& get_State();
+            std::vector<double>& get_State();
 
             //Probe activity
-            std::vector<float>& get_activity();
+            std::vector<double>& get_activity();
 
             //Probe Hebb matrix
-            std::vector<std::vector<float>>& get_hebb();
+            std::vector<std::vector<double>>& get_hebb();
 
             //Probe antiHebb matrix
-            std::vector<std::vector<float>>& get_antihebb();
+            std::vector<std::vector<double>>& get_antihebb();
 
             // Generate m_vector of projection of the current state on to stored
             //patterns
-            void generate_m(std::vector<float>& m_recipient);
+            void generate_m(std::vector<double>& m_recipient);
 
             /*---- Update methods ----*/
             //Update the activity vector of the network h(t)
@@ -127,10 +127,10 @@ namespace rate{
 
             //Update the state of the network
             void state_update_tgh(); //without input
-            void state_update_tgh(std::vector<float>& input); //with input
+            void state_update_tgh(std::vector<double>& input); //with input
 
             void state_update_rational(); //without input
-            void state_update_rational(std::vector<float>& input); //with input
+            void state_update_rational(std::vector<double>& input); //with input
 
             // IMPORTANT: both connection matrices don't considere self-connections
             // Jii = 0, in hebb_update and antiHebb_update
